feat: make add idempotent for Fake and Pypika repositories
Adding an Entities is now idempotent, and if the entity already exist in
the repository, it will update it's attributes.

fix: ignore the B608 sql injection warning

We're not letting the user define the values of the query, the only
variable inputs are the keys, that are defined by the developer, so it's
not probable that he chooses an entity attributes that are an SQL
injection. Once the #535 issue is solved, we should get rid of this
error too.