feat: make entity id definition optional
BREAKING CHANGE: We're temporarily dropping support for `str` `id_` attributes.

Right now the `Entity` class has a mandatory `id_` attribute of types `str` or
`int`. That prevents the user to create entities a model level as they are not
aware of the existent entities in the repository.

We're setting a default `id_` value of `-1` on Entities, the repository will
react to these ids, getting the last valid ID and increasing it by one.

tests: create a test entity with the mandatory name attribute

To simplify the typing of the tests.