fix: support back the string IDs
The reason of deprecation was because the new feature to auto increment the ID of
the entities that hadn't set it, wasn't "easy" to implement for strings.

In this commit we restore the support for string ids, but without the
auto increment feature. If you're using string ids is probably because
you have an id generator or you're getting them from outside sources.

doc: explain how to test applications built with repository-orm

ci: patch the flakehell extended_default_ignore error

feat: add the last and first methods to the repositories.

To return the first or last added entity to the repository.