feat: make entity models optional arguments of the repo methods
Right now the `entity_model` is a mandatory argument for these methods, that
means that the code calling the repository needs to be aware of all the entity
models inside the repository and call the methods for each of them.

Change the signatures of the `get`, `search`, `all`, `first` and `last` methods
to accept one, many or no `Entity` subclasses. If none is given, it will assume
that the repository was initialized with the `models` argument, where the user
gives the repo the model of the data it holds.

For the sake of cleanness, we're renaming `entity_models` for `models` in the
function arguments.

BREAKING: The argument `entity_model` is no longer the first one for the methods `get` and `search`, but the second.
BREAKING: The argument `database_url` is no longer the first argument in the `load_repository` function, but the second, being the first the `models`.

refactor: create the _model_not_found and _build_models abstract repo methods

chore: update requirements