feat: partially support search in attributes of type List
Allow to search for regular expressions in attributes of type
`List[str]`. FakeRepository and TinyDBRepository support it now, but
PypikaRepository doesn't, and the implementation is no trivial and I'm
in a hurry, so I'm leaving the test commented and not documenting the
feature until it is supported by all repositories.

Once the mappers are merged, this should not be difficult.