fix: remove userwarning on the add method
I've realized that it makes no sense to use it as the default method as
it shouldn't be the default behaviour for the next reasons:

* It's not intuitive when you think of `add` which can bring side
  effects to the user.
* It's beneficial when you want to partially update the existent data,
  which is kind of a corner case not the normal usage.

doc: create ADR on complex queries

Right now the `search` method only allows a dictionary of keys that the objects
must match, this way of selecting objects has become insufficient for the
programs that use the library forcing them to extract from the repository
a greater subset of objects that contain a serie of attributes and then do
another filtering afterwards. This is the case for example when the user wants
to select objects that:

* Have a datetime or integer greater or smaller than an amount.
* Don't have an attribute with a defined value
* Match a combination of criteria, for example:
    * Have an attribute with a value and another that is greater than X
    * Have an attribute with a value and another that doesn't contain the value
        Y.
* Match a criteria on a related object (`JOIN` operations).

This schema has the next disadvantages:

* The user is constrained to a very basic query type
* The repository returns more objects than the user needs, so more traffic is
    sent between the application and the database
* The user needs to do an extra filter afterwards. This is a bad idea because
    doing the filtering at databases level probably gives better performance
    and the user needs to write more code than necessary.

doc: create ADR on simplification of the search and all methods

Right now we raise an `EntityNotFoundError` exception if `search` or if `all`
returns no results, which forces the user to catch this exception when it makes
more sense to return an empty list as these methods are usually used on loops.

We're also having some issues with the types of the return values of these
methods as we allowed them to be of type `Union[Type[Entity],
List[Type[Entity]], None]` since
[003](003-make_entity_models_optional_arguments.md), both when writing the code
as the type hints are difficult to write and debug and also when doing
operations on the returned objects, as the returned list contains objects with
different models that as they have different attributes and objects they need to
be dealt differently. This last point means that the user has to split the
results using `isinstance` at some point.

Using that type in `get` is a bad idea too, as many entities may have the same
`id_`, and when you `get` you just want one entity, otherwise you'd use
`search`.