revert: add warnings on methods that don't use a single model
It's going to be deprecated as stated in [005]() in 2022-06-10.

fix: prevent search from raising exceptions when no entity is found

An argument `search_exception: bool = True` has been added to
`load_repository` and to the `Repository` `__init__` method so that
users can test what will become the default behaviour from 2022-06-10.